Moved LogicalMetricInfo to better package.
