+++ fili-core/src/main/java/com/yahoo/bard/webservice/config/LogicalMetricInfo.java
@@ -1,6 +1,8 @@
 // Copyright 2017 Yahoo Inc.
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
-package com.yahoo.bard.webservice.data.metric;
+package com.yahoo.bard.webservice.config;
+
+import com.yahoo.bard.webservice.data.metric.LogicalMetric;
 
 /**
  * Contains information passed to LogicalMetric constructor.
